+++ CHANGE_LOG.md
@@ -5,6 +5,13 @@ releases on the way from an old version to a new one.  Fix any deprecation warni
 release before upgrading to the next one.  The documentation next to each Deprecated annotation
 tells you what to use instead.  Once we delete the deprecated methods, that documentation goes too.
 
+# Release 3.1.1: (unnamed)
+
+ - Documented Cowry (Copy-On Write aRraY) better.
+  These are type-safe utility methods for dealing with arrays.
+ This project still aims to keep you from needing to deal with arrays, but sometimes you do, and it's handy to have copy-on-write routines for doing it.
+ More of these methods could be made public if they are needed.
+
 # Release 3.1.0: Map.keyIterator() and .valIterator()
  - [Nate Austin](https://github.com/navision)'s memory-efficient iteration of just keys or just values from maps.  Thank you Nate!